pyproject.toml
[tool.poetry]
name = "the-senate-discord-bot"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
flask = "^1.1.2"
replit = "^2.0.0"
discord-pretty-help = "^1.2.1"
regex = "^2021.4.4"
emoji = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
